@@ -8,17 +8,17 @@ import com.google.common.collect.Maps;
 import model.User;
 
 public class DataBase {
-    private static Map<String, User> users = Maps.newHashMap();
+  private static Map<String, User> users = Maps.newHashMap();
 
-    public static void addUser(User user) {
-        users.put(user.getUserId(), user);
-    }
+  public static void addUser(User user) {
+    users.put(user.getUserId(), user);
+  }
 
-    public static User findUserById(String userId) {
-        return users.get(userId);
-    }
+  public static User findUserById(String userId) {
+    return users.get(userId);
+  }
 
-    public static Collection<User> findAll() {
-        return users.values();
-    }
+  public static Collection<User> findAll() {
+    return users.values();
+  }
 }
