@@ -1,36 +1,44 @@
 package model;
 
 public class User {
-    private String userId;
-    private String password;
-    private String name;
-    private String email;
+  private final String password;
+  private final String userId;
+  private final String name;
+  private final String email;
 
-    public User(String userId, String password, String name, String email) {
-        this.userId = userId;
-        this.password = password;
-        this.name = name;
-        this.email = email;
-    }
+  public User(String userId, String password, String name, String email) {
+    this.userId = userId;
+    this.password = password;
+    this.name = name;
+    this.email = email;
+  }
 
-    public String getUserId() {
-        return userId;
-    }
+  public String getUserId() {
+    return userId;
+  }
 
-    public String getPassword() {
-        return password;
-    }
+  public String getPassword() {
+    return password;
+  }
 
-    public String getName() {
-        return name;
-    }
+  public String getName() {
+    return name;
+  }
 
-    public String getEmail() {
-        return email;
-    }
+  public String getEmail() {
+    return email;
+  }
 
-    @Override
-    public String toString() {
-        return "User [userId=" + userId + ", password=" + password + ", name=" + name + ", email=" + email + "]";
-    }
+  @Override
+  public String toString() {
+    return "User [userId="
+        + userId
+        + ", password="
+        + password
+        + ", name="
+        + name
+        + ", email="
+        + email
+        + "]";
+  }
 }
