@@ -10,46 +10,49 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class RequestHandler extends Thread {
-    private static final Logger log = LoggerFactory.getLogger(RequestHandler.class);
+  private static final Logger log = LoggerFactory.getLogger(RequestHandler.class);
 
-    private Socket connection;
+  private Socket connection;
 
-    public RequestHandler(Socket connectionSocket) {
-        this.connection = connectionSocket;
+  public RequestHandler(Socket connectionSocket) {
+    this.connection = connectionSocket;
+  }
+
+  public void run() {
+    log.debug(
+        "New Client Connect! Connected IP : {}, Port : {}",
+        connection.getInetAddress(),
+        connection.getPort());
+
+    try (InputStream in = connection.getInputStream();
+        OutputStream out = connection.getOutputStream()) {
+      // TODO 사용자 요청에 대한 처리는 이 곳에 구현하면 된다.
+      DataOutputStream dos = new DataOutputStream(out);
+      byte[] body = "Hello World".getBytes();
+      response200Header(dos, body.length);
+      responseBody(dos, body);
+    } catch (IOException e) {
+      log.error(e.getMessage());
     }
+  }
 
-    public void run() {
-        log.debug("New Client Connect! Connected IP : {}, Port : {}", connection.getInetAddress(),
-                connection.getPort());
-
-        try (InputStream in = connection.getInputStream(); OutputStream out = connection.getOutputStream()) {
-            // TODO 사용자 요청에 대한 처리는 이 곳에 구현하면 된다.
-            DataOutputStream dos = new DataOutputStream(out);
-            byte[] body = "Hello World".getBytes();
-            response200Header(dos, body.length);
-            responseBody(dos, body);
-        } catch (IOException e) {
-            log.error(e.getMessage());
-        }
+  private void response200Header(DataOutputStream dos, int lengthOfBodyContent) {
+    try {
+      dos.writeBytes("HTTP/1.1 200 OK \r\n");
+      dos.writeBytes("Content-Type: text/html;charset=utf-8\r\n");
+      dos.writeBytes("Content-Length: " + lengthOfBodyContent + "\r\n");
+      dos.writeBytes("\r\n");
+    } catch (IOException e) {
+      log.error(e.getMessage());
     }
+  }
 
-    private void response200Header(DataOutputStream dos, int lengthOfBodyContent) {
-        try {
-            dos.writeBytes("HTTP/1.1 200 OK \r\n");
-            dos.writeBytes("Content-Type: text/html;charset=utf-8\r\n");
-            dos.writeBytes("Content-Length: " + lengthOfBodyContent + "\r\n");
-            dos.writeBytes("\r\n");
-        } catch (IOException e) {
-            log.error(e.getMessage());
-        }
+  private void responseBody(DataOutputStream dos, byte[] body) {
+    try {
+      dos.write(body, 0, body.length);
+      dos.flush();
+    } catch (IOException e) {
+      log.error(e.getMessage());
     }
-
-    private void responseBody(DataOutputStream dos, byte[] body) {
-        try {
-            dos.write(body, 0, body.length);
-            dos.flush();
-        } catch (IOException e) {
-            log.error(e.getMessage());
-        }
-    }
+  }
 }
